Implement Akka actor system for image submission processing.

+++ app/utils/AwsS3Utils.java
@@ -1,5 +1,7 @@
 package utils;
 
+import akka.dispatch.Futures;
+import com.amazonaws.AmazonClientException;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.event.ProgressEvent;
@@ -9,7 +11,10 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.transfer.Copy;
 import com.amazonaws.services.s3.transfer.TransferManager;
 import com.amazonaws.services.s3.transfer.Upload;
+import play.Logger;
 import play.Play;
+import scala.concurrent.Future;
+import scala.concurrent.Promise;
 
 import java.io.*;
 import java.nio.file.Files;
@@ -19,7 +24,7 @@ import java.nio.file.StandardCopyOption;
 public class AwsS3Utils {
 
     /**
-     * Utility method to upload file to S3.
+     * Utility method to upload file to S3. Note that this method blocks until the upload has completed or failed.
      *
      * @param file The file to upload.
      * @return
@@ -40,15 +45,15 @@ public class AwsS3Utils {
         upload.addProgressListener(new ProgressListener() {
             @Override
             public void progressChanged(ProgressEvent progressEvent) {
-                System.out.print("\rS3 Upload Progress (" + file.getName() + "): " + upload.getProgress().getPercentTransferred() + "%");
+                Logger.debug("\rS3 Upload Progress (" + file.getName() + "): " + upload.getProgress().getPercentTransferred() + "%");
 
                 switch(progressEvent.getEventType()) {
                     case TRANSFER_COMPLETED_EVENT:
-                        System.out.println("\n" + file.getName() + " has been successfully uploaded to S3.");
+                        Logger.debug("\n" + file.getName() + " has been successfully uploaded to S3.");
                         break;
                     case TRANSFER_FAILED_EVENT:
-                        System.out.println("\n" + file.getName() + " upload failed.");
-
+                        Logger.debug("\n" + file.getName() + " upload failed.");
+                        break;
                 }
             }
         });
@@ -70,6 +75,7 @@ public class AwsS3Utils {
 
     /**
      * Utility method to upload file to S3 with given file name and contentType metadata.
+     * Note that this method blocks until upload has completed or failed.
      *
      * @param file The file to upload.
      * @param filename The real file name, which is also used as the S3 object key name.
@@ -124,6 +130,60 @@ public class AwsS3Utils {
         }
     }
 
+    /**
+     * Utility method to upload file to S3. Method is non-blocking and returns a Future.
+     *
+     * @param file The file to upload.
+     * @return A Future holding the resulting s3 url for the uploaded file, or an AmazonClientException on failure.
+     */
+    public static Future<String> uploadFileAsync(final File file) {
+        // Create empty promise, which will hold the upload status. We will return its future.
+        final Promise<String> promise = Futures.promise();
+
+        final String awsAccessKey = Play.application().configuration().getString("aws.access.key");
+        final String awsSecretKey = Play.application().configuration().getString("aws.secret.key");
+        final String s3BucketName = "starmapper-app";
+
+        final AWSCredentials awsCred = new BasicAWSCredentials(awsAccessKey, awsSecretKey);
+
+        final TransferManager tm = new TransferManager(awsCred);
+
+        //final Upload upload = tm.upload(s3BucketName, filename, file);
+        final Upload upload = tm.upload(s3BucketName, file.getName(), file);
+
+        // Create and set anonymous ProgressListener
+        upload.addProgressListener(new ProgressListener() {
+            @Override
+            public void progressChanged(ProgressEvent progressEvent) {
+                //System.out.print("\rS3 Upload Progress (" + file.getName() + "): " + upload.getProgress().getPercentTransferred() + "%");
+
+                switch(progressEvent.getEventType()) {
+                    case TRANSFER_COMPLETED_EVENT:
+                        // Complete the promise with the newly created s3 url for the file.
+                        //Logger.debug(file.getName() + " has been successfully uploaded to S3.");
+                        promise.success("https://s3.amazonaws.com/" + s3BucketName + "/" + file.getName());
+                        break;
+                    case TRANSFER_FAILED_EVENT:
+                        //Logger.debug(file.getName() + " upload failed.");
+
+                        try {
+                            // Shouldn't block because transfer already finished at this point.
+                            AmazonClientException ex = upload.waitForException();
+                            // Complete the promise with the exception.
+                            promise.failure(ex);
+                        } catch (InterruptedException e) {
+                            e.printStackTrace();
+                        }
+
+                        break;
+                }
+            }
+        });
+
+        // Return the future associated with the promise.
+        return promise.future();
+    }
+
     public static void setMetadata(ObjectMetadata metadata, String bucket, String key) {
         String awsAccessKey = Play.application().configuration().getString("aws.access.key");
         String awsSecretKey = Play.application().configuration().getString("aws.secret.key");
@@ -163,17 +223,21 @@ public class AwsS3Utils {
         // Get Path representation of File.
         Path path = file.toPath();
 
-        try {
-            // Renaming is done via platform dependent move command.
-            // We opt for Files.move() rather than File.renameTo() because the former seems to be more reliable on
-            // different platforms.
-            Path newPath = Files.move(path, path.resolveSibling(newFileName), StandardCopyOption.REPLACE_EXISTING);
-            File newFile = newPath.toFile();
-            newFile.deleteOnExit();
+        // Check to see that the new file name has not already been set.
+        if (!path.getFileName().equals(newFileName)) {
+            try {
+                // Renaming is done via platform dependent move command.
+                // We opt for Files.move() rather than File.renameTo() because the former seems to be more reliable on
+                // different platforms.
+                // Note: Changed to copy() command so that we can hold on to original file.
+                Path newPath = Files.copy(path, path.resolveSibling(newFileName), StandardCopyOption.REPLACE_EXISTING);
+                File newFile = newPath.toFile();
+                newFile.deleteOnExit();
 
-            return newFile;
-        } catch (IOException e) {
-            e.printStackTrace();
+                return newFile;
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
 
         return null;
